[AC-2361] Refactor StripeController (#4136)
* Changes ensures provider_id is handled and stored for Braintree.


* refactoring of the stripeController class


* Move the constant variables to utility class


* Adding comments to the methods


* Add more comments to describe the method


* Add the providerId changes


* Add the missing providerId


* Fix the IsSponsoredSubscription bug


---------
