+++ src/Billing/Startup.cs
@@ -52,6 +52,21 @@ public class Startup
         // Context
         services.AddScoped<ICurrentContext, CurrentContext>();
 
+        //Handlers
+        services.AddScoped<IStripeEventUtilityService, StripeEventUtilityService>();
+        services.AddScoped<ISubscriptionDeletedHandler, SubscriptionDeletedHandler>();
+        services.AddScoped<ISubscriptionUpdatedHandler, SubscriptionUpdatedHandler>();
+        services.AddScoped<IUpcomingInvoiceHandler, UpcomingInvoiceHandler>();
+        services.AddScoped<IChargeSucceededHandler, ChargeSucceededHandler>();
+        services.AddScoped<IChargeRefundedHandler, ChargeRefundedHandler>();
+        services.AddScoped<ICustomerUpdatedHandler, CustomerUpdatedHandler>();
+        services.AddScoped<IInvoiceCreatedHandler, InvoiceCreatedHandler>();
+        services.AddScoped<IPaymentFailedHandler, PaymentFailedHandler>();
+        services.AddScoped<IPaymentMethodAttachedHandler, PaymentMethodAttachedHandler>();
+        services.AddScoped<IPaymentSucceededHandler, PaymentSucceededHandler>();
+        services.AddScoped<IInvoiceFinalizedHandler, InvoiceFinalizedHandler>();
+        services.AddScoped<IStripeEventProcessor, StripeEventProcessor>();
+
         // Identity
         services.AddCustomIdentityServices(globalSettings);
         //services.AddPasswordlessIdentityServices<ReadOnlyDatabaseIdentityUserStore>(globalSettings);
